refactor: deleting unnecessary data
