+++ src/app/page.tsx
@@ -1,21 +1,9 @@
 "use client";
 import HeroContent from "@/components/HeroContent";
 import heroData from "./../data/heroData.json";
-import BookingPopup from "@/components/PopUp/Booking";
-import CheckoutPopup from "@/components/PopUp/Checkout";
-import { Button } from "@/components/ui/Button";
-import { useState } from "react";
-import CardDetailsPopup from "@/components/PopUp/CardDetails";
+import PopUp from "@/components/PopUp";
 
 export default function Home() {
-  const [showBooking, setShowBooking] = useState(false);
-  const [showCheckout, setShowCheckout] = useState(false);
-  const [showCardDetails, setShowCardDetails] = useState(false);
-
-  const handleProceedToCheckout = () => {
-    setShowBooking(false);
-    setTimeout(() => setShowCheckout(true), 200);
-  };
   return (
     <div>
       <div className="flex items-center justify-center">
@@ -28,29 +16,7 @@ export default function Home() {
         imageUrl={heroData.imageUrl}
         content={{ title: heroData.title, description: heroData.description }}
       />
-      {/* <PopUp /> */}
-      <div>
-        <div className="flex justify-center mt-10">
-          <Button variant="default" onClick={() => setShowBooking(true)}>
-            Book Now
-          </Button>
-        </div>
-        {showBooking && (
-          <BookingPopup
-            onProceed={handleProceedToCheckout}
-            onClose={() => setShowBooking(false)}
-          />
-        )}
-        {showCheckout && (
-          <CheckoutPopup
-            onClose={() => setShowCheckout(false)}
-            onSelectCard={() => setShowCardDetails(true)}
-          />
-        )}
-        {showCardDetails && (
-          <CardDetailsPopup onClose={() => setShowCardDetails(false)} />
-        )}
-      </div>
+      <PopUp />
     </div>
   );
 }
